+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Compatibility with TYPO3 v11 (#16)
 - Compatibility with TYPO3 v12 (#28)
+- Support of Matomo v4
 
 ## [2.4.0] - 2025-09-17