+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0] - 2025-11-26
+
 ### Added
 - Compatibility with TYPO3 v14
 
@@ -125,7 +127,8 @@ Documentation improvements
 Initial release
 
 
-[Unreleased]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v2.4.0...HEAD
+[Unreleased]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v3.0.0...HEAD
+[3.0.0]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v2.4.0...v3.0.0
 [2.4.0]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v2.3.0...v2.4.0
 [2.3.0]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v2.1.1...v2.2.0